FatFs can work in debug mode, but there are two problem in new version. 1. In release version, FatFs can not work. 2. SD card DMA transfer has problem in one block writing, and DMA mode can return right now for FatFs.

+++ rtos/TEST_FreeRTOS/Driver/src/stm32f746g_disco.c
@@ -730,7 +730,7 @@ bool_t checkDMA(uint16_t data)
 	return (*pDes == data && *(pDes + 0x3FF8) == data);
 }
 
-bool_t setSDCardData(const uint16_t blockAddr)
+bool_t setSDCardData(const uint32_t blockAddr)
 {
 	// Set source data
 	for (uint16_t i = 0; i < SD_BLOCK_SIZE * 2; i++) {
@@ -744,7 +744,7 @@ bool_t setSDCardData(const uint16_t blockAddr)
 	return True;
 }
 
-bool_t getSDCardData(const uint16_t blockAddr)
+bool_t getSDCardData(const uint32_t blockAddr)
 {
 	// Read back data from SD card
 	//if (!sdPollingRead(blockAddr, 2, &sdTestDes[0])) return False;
